requirements.txt
django == 4.2.2
psycopg2-binary
python-dotenv == 1.0.0
pre-commit==3.3.3
django-debug-toolbar==4.1.0
django-material-admin==1.8.6
django-redis-cache==3.0.1
djangorestframework==3.14.0
markdown ==3.3.6
django-filter==23.2
django-crispy-forms==2.0
xhtml2pdf==0.2.11
firebase-admin==6.2.0
sentry-sdk==1.29.0
django-webpush==0.3.5
weasyprint==59.0
gunicorn==21.2.0
boto3==1.26.59
django-storages==1.13.2
django-allauth==0.54.0
pytest==7.4.0
pytest-django==4.5.2    
factory-boy==3.3.0
coverage==7.3.0
django-webtest==1.9.10
django-anymail
django_mailgun_mime
factory_boy==3.3.0